Remove personal access token from git repo
Storing a different personal access token on each computer so that they
can be separately revokable, so don't want it tracked in git. (Also it's
more secure to just keep the access tokens on each computer so they
never leave the computer.)

Once entered into git for the first time (e.g. the first time you try to
clone a private repo over HTTPS) the access token or password will be
stored on the machine in ~/.git-credentials.